Pull argument primitive fields out into their own files
